feat: implement MobileLoadingSkeleton and useInfiniteScroll for improved loading experience and infinite scrolling
